pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kifrs-agent"
version = "0.1.0"
description = "K-IFRS 질의응답 Agent — LangGraph + DeepAgents 기반"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.12",
    "langgraph>=1.1.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-anthropic>=1.3.25",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.2.0",
    "pgvector>=0.3.0",
    "openai>=1.0.0",
    "cohere>=5.0.0",
    "kiwipiepy>=0.18.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
    "langgraph-cli>=0.4.18",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4"]

[tool.ruff.lint.per-file-ignores]
"app/prompts.py" = ["E501"]  # 프롬프트 문자열 내부 줄 길이 제외
"eval/evaluate.py" = ["E402"]  # __main__ 스크립트: load_dotenv() 후 import
"app/migrations/*.py" = ["E402"]  # 마이그레이션 스크립트: sys.path 조정 후 import

[tool.pytest.ini_options]
testpaths = ["tests"]
